Steps taken in order to compile with c++11

+++ src/main.cpp
@@ -99,7 +99,7 @@ int main(int argc, char *argv[])
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x};
+            static_cast<GLuint>(face->glyph->advance.x)};
         Characters.insert(std::pair<GLchar, Character>(c, character));
     }
     glBindTexture(GL_TEXTURE_2D, 0);
